requirements.txt
fastapi
uvicorn
torch
torchvision
pandas
scikit-learn
faiss-cpu
pillow
requests
multipart
